Revert "Merge branch 'demo' into recette"
This reverts commit 4ae25b5e26b805d7aeac3cb8e827c9bbf1224929.

+++ back/src/main/java/com/ecoassitant/back/dto/ConstanteDto.java
@@ -13,7 +13,7 @@ import java.util.Objects;
 @AllArgsConstructor
 public class ConstanteDto {
     private Long idConstante;
-    private double constante;
+    private int constante;
     private String tracabilite;
 
     /**
